Extract message processing to class
